+++ demisto_sdk/commands/validate/validate_manager.py
@@ -672,7 +672,8 @@ class ValidateManager:
                                                                validate_dependencies=not self.skip_dependencies,
                                                                id_set_path=id_set_path,
                                                                private_repo=self.is_external_repo,
-                                                               skip_id_set_creation=self.skip_id_set_creation)
+                                                               skip_id_set_creation=self.skip_id_set_creation,
+                                                               prev_ver=self.prev_ver)
         pack_errors = pack_unique_files_validator.validate_pack_unique_files()
         if pack_errors:
             click.secho(pack_errors, fg="bright_red")